remove detritus before building
Issue #102 was happening because the database was giving instructions to
destroy the file that was being built, but the order of operations was
to clean up AFTER building, which ended up destroying files that were
just built. This will fix #102 by flipping around the order.

NOTE: I should add crayon indicator here that tells people WHAT files
are being built/copied.