fix issue with disappearing remote
Surprise surprise: it turns out the stackoverflow solution worked for
what I needed at the time, but was certainly not a general solution.
Here, I'm spelling out the explicit refspec string so that git will know
which branch I'm fetching and _hopefully_ it will be more clear for
future folk.