update tests for branch presence
The CI deploy tests were giving strange results on different systems, so
it was better to test the _concept_ of a branch _existing_ rather than
the number of commits that were recorded. I believe the issue with the
differing number of commits on the SITE branch is due to time splitting
between minutes across tests (e.g. if the first deploy test happens in
the same minute as the second deploy test, then there would be no
change, but if it happened across times, then there would be three
commits).