add instructor page building
N.B. this does slow things down a bit, unfortunately. The upside is that
we are no longer rendering the time modified in the metadata (that is,
in the last-modified meta element in the head), so this should help
avoid un-necessary rewrites.

This will hopefully change if we can figure out a better way to provide
a JS solution with redirects.