fix tests; setup for refactor
This is getting ridiculous. The tests now take > 5 minutes to run on my
machine and it's clear that the bottleneck for us is the fact that we
check every HTML page with pkgdown regardless of whether or not it needs
updating.

I've set this up for a refactor, but until that happens it's gonna be
slow.