always provision learners/setup.md
This ensures that the setup section is always provisioned even if it
does not appear in the config list.

This also adds a test which explicitly sets up the learner config to a
single markdown file and ensures that the setup.md is still built.

This will fix https://github.com/carpentries/sandpaper/issues/386