update root criteria
This will address https://github.com/carpentries/workbench/issues/65
by adding fallbacks to the episodes directory criteria. I've added the
other directories we expect so that when the episodes directory does not
exist (in the case of overview pages), then we can check that the site
directory or any of the other directories exists.

I had initially attempted this by setting the config.yaml as the
criteria, but I forgot that I had also copied the config.yaml over to
the site/built folder, so there were a lot of failures from this one
change because it was getting confused where the root of the project.

Another tactic I considered was to check if `.git/` was a directory, but
then I realized this wouldn't work for submodules.