add test for fixing overview metadata context
It turns out that if there is _extra_ metadata, it will be carried over
into the next lesson as the metadata is not cleared. This puts a patch
for the most important ones, but I am going to need to include something
that clears out the old metadata within the lesson validation process.