copy overview assets if they exist
This was a manual process in build_markdown() that was taking valuable
mental real-estate and it was frustrating to read because when I would
look at it, it didn't immediately tell me what it was doing (copy all
assets related to the build including non-markdown files and folders).

With the addition of the overview pages, it was time to move this into a
function so that I could have a bit more room to be verbose about what I
was doing without distracting from build_markdown().