set names rel path names for child list
I was getting in trouble because the order of the files from pegboard
and sandpaper's file listing system do not match (because sandpaper
cares about more files).

The bed that I made for pegboard initially set the names of the files to
be the file names without the paths and that was it, but in the
database, the IDs for the files were the names of the files relative to
the lesson path and thus it caused confusion. Fixing the relative names
solved this problem (and likely I should adopt this for pegboard itself
:/)