reset metdata for handout test
This is one of the weird litte things that I need to get better at
controlling. When we test these things, the global metdata is affected,
but it's not possible (at the moment) to _remove_ metadata without
resetting everything, so instead, we need to set the global metadata
option for "handout" to NULL.

In the words of the author of a C++ library that I rewrote in R for my
Ph. D.:

> This is a bloody awful nasty hack ... and will give us grief
> elsewhere. Find a better way to do this