allow translated vars to be included in site data
The reason why the initial apporach I proposed was not working is
because this bumps into a few concepts that have not been clearly
documented on this end.

First, there is the concept of the `pkg` object. This object is a
representation of the _pkgdown.yaml file from the `site/` folder. When
any page is built, this is passed to `pkgdown::data_template()`, which
creates the data needed for the whisker templates (e.g. everything in
the `$params` section of the `_pkgdown.yaml` file is shunted into `$yaml`)
and boilerplate items like dropdown menu items (for pkgdown) are added.

In my original conception, I would be able to add `$translate` to the
`pkg` object and it would be propogated to the website. This was not
true and this is where the disconnect initially lies. The problem was that the
code for `pkgdown::data_template()` was _replacing_ the `$translate`
field with the default translations from {pkgdown}.

In this package, the way we pass variables to {varnish} via the
`_pkgdown.yaml` file, the generated content for `learner_globals()` and
`instructor_globals()`, and the `this_metadata()` object, which contains
both supplied and generated metadata content.

All of these are passed to `build_html()`, which uses
`pkgdown::render_page_html()`, passing the generated global data to the
`data` parameter and _this_ is where the translations are incorporated.