add mechanism to include HTML templating
There are _many_ situations where we want to translate something that
has a link _inside_ of the translation itself. It's rather awkward to
force the translator to translate something like this while keeping the
HTML:

Please visit <a href="CODE_OF_CONDUCT.html">the Code of Conduct</a>

Because we know what these HTML elements are supposed to be, we can get
away with this by using <( and )> for the start and end pieces of HTML
that we wrap.