collect translations; add my attempt for SpanToTop
The Back To Top link is split out so that the word "Back" is encased in
a `<span>` element that disappears on smaller screens so that the
message becomes "To Top".

This update translates `<(Back)> To Top` in a way that allows the translator to
choose which text should disappear on smaller screens. In languages with
SVO syntax (as English is), the order will stay the same and the word
for "Back" will be on the top, but for languages with SOV syntax
(Japanese, Korean, German), it will appear as `To Top <(Back)>`. I've
attempted to do this for the Japanese translation, but it probably
should be corrected.

I'm not sure how flexible this is, but this is an attempt.