replace translate_callout_heading
This replaces it with a more efficient method by translating once and
then applying those translations over the nodes. This is still not
efficient as can be. The translations _could_ live in yet another global
environment that can be initialized when the config file is read in, but
we will optimize that later