remove global var translations for aggregators
One of the problems that lead me to 7d9cc2cc6ff81b78efd31acfce00a2b8d34de81c
was the fact that a test for `build_lesson()` was failing in the
situation where we removed the instructor notes page from the source and
rebuilt it using the instructor note aggregator function without the
`build_lesson()` context.

When `this_lesson()` was called, it clobbered the `$translation` items
and then we got an error in that test because the translation for the
title was not available.

I will reiterate that this problem is almost non-existent because none
of the HTML generating functions are exported nor are they called
directly by users, but it still bothers me. In fact, it would be a good
idea to centralise the translation to a single function and always
assume that the global data will always exist. This would help avoid
confusing errors in the language of the lesson (as opposed to the
language of the user).